code ref/i7 OCR/requirements.txt
streamlit==1.20.0
streamlit_option_menu==0.3.2
pymupdf==1.21.1
streamlit_javascript==0.1.5
python-dotenv==1.0.0